+++ src/common/types/apiResponse.ts
@@ -1,5 +1,6 @@
 export interface apiResponse<T> {
   data: T;
+  success?: boolean;
   status: number;
   message: string;
 }
