+++ src/my/hooks/useMyPage.tsx
@@ -1,5 +1,8 @@
 import { useState, useRef, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
+import { useToast } from "../../common/hooks/useToastMsg";
+import { getUserService } from "../services/getUserService";
+import { updateAccountNumberService } from "../services/updateAccountNumber";
 
 // 타입 정의
 export interface ConfirmContentType {
@@ -10,13 +13,14 @@ export interface ConfirmContentType {
 
 export const useMyPage = () => {
   const navigate = useNavigate();
+  const { showToast, isToastVisible, toastMessage } = useToast();
 
   // 계좌 관련 상태
   const [isEditing, setIsEditing] = useState(false);
-  const [bankName, setBankName] = useState(
-    "국민은행아아아아아아아아아아아아아아아아"
-  );
-  const [accountNumber, setAccountNumber] = useState("1234567890");
+  const [userName, setUserName] = useState("");
+  const [bankName, setBankName] = useState("");
+  const [accountNumber, setAccountNumber] = useState("");
+  const [profileImg, setProfileImg] = useState("");
 
   // 참조
   const bankInputRef = useRef<HTMLInputElement>(null);
@@ -27,6 +31,30 @@ export const useMyPage = () => {
   const [confirmContent, setConfirmContent] =
     useState<ConfirmContentType | null>(null);
 
+  // 유저 정보 조회
+  useEffect(() => {
+    const fetchUserData = async () => {
+      try {
+        const token = localStorage.getItem("accessToken");
+        if (!token) {
+          showToast("로그인이 필요합니다.");
+          return;
+        }
+
+        const res = await getUserService(token);
+
+        setUserName(res.data.userName);
+        setBankName(res.data.bankName);
+        setAccountNumber(res.data.accountNumber);
+        setProfileImg(res.data.profileImageUrl);
+      } catch {
+        showToast("유저 조회 중 오류가 발생했어요.");
+      }
+    };
+
+    fetchUserData();
+  }, []);
+
   // 수정 모드 포커스 효과
   useEffect(() => {
     if (isEditing && accountInputRef.current) {
@@ -35,10 +63,33 @@ export const useMyPage = () => {
   }, [isEditing]);
 
   // 계좌 수정 토글
-  const handleEditToggle = () => {
+  const handleEditToggle = async () => {
     if (isEditing) {
-      // Save logic would go here
+      const token = localStorage.getItem("accessToken");
+      if (!token) {
+        showToast("로그인이 필요합니다.");
+        return;
+      }
+
+      const requestData = {
+        bankName: bankName,
+        accountNumber: accountNumber,
+      };
+
+      try {
+        const res = await updateAccountNumberService(requestData, token);
+
+        if (res.success == true) {
+          showToast("계좌 정보가 수정되었어요.");
+        } else {
+          const errorMessage = res.message || "계좌 수정에 실패했어요.";
+          showToast(errorMessage);
+        }
+      } catch {
+        showToast("계좌 정보 수정 중 오류가 발생했어요.");
+      }
     }
+
     setIsEditing(!isEditing);
   };
 
@@ -60,6 +111,7 @@ export const useMyPage = () => {
 
   // 로그아웃 확인 처리
   const handleConfirm = () => {
+    localStorage.removeItem("accessToken");
     navigate("/login");
     closeConfirmModal();
   };
@@ -70,8 +122,10 @@ export const useMyPage = () => {
   return {
     // 상태
     isEditing,
+    userName,
     bankName,
     accountNumber,
+    profileImg,
     showConfirmModal,
     confirmContent,
 
@@ -87,5 +141,9 @@ export const useMyPage = () => {
     closeConfirmModal,
     handleConfirm,
     goToTerms,
+
+    // 토스트 메세지
+    isToastVisible,
+    toastMessage,
   };
 };
