+++ src/my/styles/myPageStyles.ts
@@ -1,0 +1,19 @@
+/**
+ * 마이페이지 커스텀 스타일
+ * @author 희진
+ */
+
+export const myPageStyles = {
+  text: {
+    title2: "font-pretendard font-normal text-Title2",
+    body1: "font-pretendard font-normal text-Body1",
+    body2: "font-pretendard font-normal text-Body2",
+    body3: "font-pretendard font-normal text-Body3",
+  },
+  color: {
+    black: "text-black",
+    gray40: "text-gray40",
+    gray78: "text-gray78",
+    lightGray: "text-grayC8",
+  },
+};
