+++ src/my/types/getUserResponse.ts
@@ -1,0 +1,7 @@
+export interface getUserResponse {
+  userId: number;
+  userName: string;
+  profileImageUrl: string;
+  bankName: string;
+  accountNumber: string;
+}
