+++ src/my/types/updateAccountNumberRequest.ts
@@ -1,0 +1,4 @@
+export interface updateAccountNumberRequest {
+  bankName: string;
+  accountNumber: string;
+}
