+++ src/my/types/updateAccountNumberResponse.ts
@@ -1,0 +1,3 @@
+export interface updateAccountNumberResponse {
+  success: boolean;
+}
